requirements.txt
asgiref==3.5.1
certifi==2021.5.30
charset-normalizer==2.0.12
distro==1.6.0
dj-database-url==0.5.0
django-embed-video==1.4.4
django-environ==0.4.5
django-tawkto==0.3
django-widget-tweaks==1.4.8
djangoembed==0.1.1
idna==3.3
jws==0.1.3
msgpack==1.0.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
PyMySQL==1.0.2
pyparsing==2.4.7
python-decouple==3.4
python-jwt==2.0.1
pytz==2021.1
requests==2.27.1
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
validate-email==1.3
whitenoise==5.2.0
